auth-service/requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
pyjwt==2.9.0
werkzeug==3.0.3
